+++ test/CosmosDB.collections.Tests.ps1
@@ -14,7 +14,7 @@ InModuleScope CosmosDB {
     $script:testDatabase = 'testDatabase'
     $script:testKey = 'GFJqJeri2Rq910E0G7PsWoZkzowzbj23Sm9DUWFC0l0P8o16mYyuaZKN00Nbtj9F1QQnumzZKSGZwknXGERrlA=='
     $script:testKeySecureString = ConvertTo-SecureString -String $script:testKey -AsPlainText -Force
-    $script:testConnection = [CosmosDb.Connection] @{
+    $script:testContext = [CosmosDb.Context] @{
         Account  = $script:testAccount
         Database = $script:testDatabase
         Key      = $script:testKeySecureString
@@ -66,6 +66,112 @@ InModuleScope CosmosDB {
         }
     }
 
+    Describe 'New-CosmosDbCollectionIncludedPathIndex' -Tag 'Unit' {
+        It 'Should exist' {
+            { Get-Command -Name New-CosmosDbCollectionIncludedPathIndex  -ErrorAction Stop } | Should -Not -Throw
+        }
+
+        Context 'Called with all parameters' {
+            $script:result = $null
+
+            It 'Should not throw exception' {
+                $newCosmosDbCollectionIncludedPathIndexParameters = @{
+                    Kind      = 'Hash'
+                    DataType  = 'String'
+                    Precision = -1
+                }
+
+                { $script:result = New-CosmosDbCollectionIncludedPathIndex @newCosmosDbCollectionIncludedPathIndexParameters } | Should -Not -Throw
+            }
+
+            It 'Should return expected result' {
+                $script:result | Should -BeOfType 'CosmosDB.IndexingPolicy.Path.Index'
+                $script:result.Kind | Should -Be 'Hash'
+                $script:result.DataType | Should -Be 'String'
+                $script:result.Precision | Should -Be -1
+            }
+        }
+    }
+
+    Describe 'New-CosmosDbCollectionIncludedPath' -Tag 'Unit' {
+        It 'Should exist' {
+            { Get-Command -Name New-CosmosDbCollectionIncludedPath  -ErrorAction Stop } | Should -Not -Throw
+        }
+
+        Context 'Called with all parameters' {
+            $script:result = $null
+
+            It 'Should not throw exception' {
+                $newCosmosDbCollectionIncludedPathParameters = @{
+                    Path  = '/*'
+                    Index = (New-CosmosDbCollectionIncludedPathIndex -Kind 'Hash' -DataType 'String' -Precision -1)
+                }
+
+                { $script:result = New-CosmosDbCollectionIncludedPath @newCosmosDbCollectionIncludedPathParameters } | Should -Not -Throw
+            }
+
+            It 'Should return expected result' {
+                $script:result | Should -BeOfType 'CosmosDB.IndexingPolicy.Path.IncludedPath'
+                $script:result.Path | Should -Be '/*'
+                $script:result.Indexes[0].Kind | Should -Be 'Hash'
+                $script:result.Indexes[0].DataType | Should -Be 'String'
+                $script:result.Indexes[0].Precision | Should -Be -1
+            }
+        }
+    }
+
+    Describe 'New-CosmosDbCollectionExcludedPath' -Tag 'Unit' {
+        It 'Should exist' {
+            { Get-Command -Name New-CosmosDbCollectionExcludedPath  -ErrorAction Stop } | Should -Not -Throw
+        }
+
+        Context 'Called with all parameters' {
+            $script:result = $null
+
+            It 'Should not throw exception' {
+                $newCosmosDbCollectionExcludedPathParameters = @{
+                    Path  = '/*'
+                }
+
+                { $script:result = New-CosmosDbCollectionExcludedPath @newCosmosDbCollectionExcludedPathParameters } | Should -Not -Throw
+            }
+
+            It 'Should return expected result' {
+                $script:result | Should -BeOfType 'CosmosDB.IndexingPolicy.Path.ExcludedPath'
+                $script:result.Path | Should -Be '/*'
+            }
+        }
+    }
+
+    Describe 'New-CosmosDbCollectionIndexingPolicy' -Tag 'Unit' {
+        It 'Should exist' {
+            { Get-Command -Name New-CosmosDbCollectionIndexingPolicy  -ErrorAction Stop } | Should -Not -Throw
+        }
+
+        Context 'Called with all parameters' {
+            $script:result = $null
+
+            It 'Should not throw exception' {
+                $newCosmosDbCollectionIndexingPolicyParameters = @{
+                    Automatic = $true
+                    IndexingMode = 'Consistent'
+                    IncludedPath = (New-CosmosDbCollectionIncludedPath -Path '/*' -Index (New-CosmosDbCollectionIncludedPathIndex -Kind 'Hash' -DataType 'String' -Precision -1))
+                    ExcludedPath = (New-CosmosDbCollectionExcludedPath -Path '/*')
+                }
+
+                { $script:result = New-CosmosDbCollectionIndexingPolicy @newCosmosDbCollectionIndexingPolicyParameters } | Should -Not -Throw
+            }
+
+            It 'Should return expected result' {
+                $script:result | Should -BeOfType 'CosmosDB.IndexingPolicy.Policy'
+                $script:result.Automatic | Should -Be $true
+                $script:result.IndexingMode | Should -Be 'Consistent'
+                $script:result.IncludedPaths[0].Path | Should -Be '/*'
+                $script:result.ExcludedPaths[0].Path | Should -Be '/*'
+            }
+        }
+    }
+
     Describe 'Get-CosmosDbCollection' -Tag 'Unit' {
         It 'Should exist' {
             { Get-Command -Name Get-CosmosDbCollection -ErrorAction Stop } | Should -Not -Throw
@@ -75,7 +181,7 @@ InModuleScope CosmosDB {
             Mock -CommandName Set-CosmosDbCollectionType -MockWith { $Collection }
         }
 
-        Context 'Called with connection parameter and no Id' {
+        Context 'Called with context parameter and no Id' {
             $script:result = $null
 
             Mock `
@@ -85,7 +191,7 @@ InModuleScope CosmosDB {
 
             It 'Should not throw exception' {
                 $getCosmosDbCollectionParameters = @{
-                    Connection = $script:testConnection
+                    Context = $script:testContext
                 }
 
                 { $script:result = Get-CosmosDbCollection @getCosmosDbCollectionParameters } | Should -Not -Throw
@@ -105,7 +211,7 @@ InModuleScope CosmosDB {
             }
         }
 
-        Context 'Called with connection parameter and an Id' {
+        Context 'Called with context parameter and an Id' {
             $script:result = $null
 
             Mock `
@@ -115,8 +221,8 @@ InModuleScope CosmosDB {
 
             It 'Should not throw exception' {
                 $getCosmosDbCollectionParameters = @{
-                    Connection = $script:testConnection
-                    Id         = $script:testCollection1
+                    Context = $script:testContext
+                    Id      = $script:testCollection1
                 }
 
                 { $script:result = Get-CosmosDbCollection @getCosmosDbCollectionParameters } | Should -Not -Throw
@@ -144,18 +250,23 @@ InModuleScope CosmosDB {
             Mock -CommandName Set-CosmosDbCollectionType -MockWith { $Collection }
         }
 
-        Context 'Called with connection parameter and an Id' {
+        Context 'Called with context parameter and an Id' {
+            $invokecosmosdbrequest_parameterfilter = {
+                $Method -eq 'Post' -and `
+                    $ResourceType -eq 'colls' -and `
+                    $Body -eq (ConvertTo-Json -Depth 10 -InputObject @{ id = $script:testCollection1 } )
+            }
             $script:result = $null
 
             Mock `
                 -CommandName Invoke-CosmosDbRequest `
-                -ParameterFilter { $Method -eq 'Post' -and $ResourceType -eq 'colls' -and $Body -eq "{ `"id`": `"$($script:testCollection1)`" }" } `
+                -ParameterFilter $invokecosmosdbrequest_parameterfilter `
                 -MockWith { ConvertFrom-Json -InputObject $script:testJsonSingle }
 
             It 'Should not throw exception' {
                 $newCosmosDbCollectionParameters = @{
-                    Connection = $script:testConnection
-                    Id         = $script:testCollection1
+                    Context = $script:testContext
+                    Id      = $script:testCollection1
                 }
 
                 { $script:result = New-CosmosDbCollection @newCosmosDbCollectionParameters } | Should -Not -Throw
@@ -168,22 +279,28 @@ InModuleScope CosmosDB {
             It 'Should call expected mocks' {
                 Assert-MockCalled `
                     -CommandName Invoke-CosmosDbRequest `
-                    -ParameterFilter { $Method -eq 'Post' -and $ResourceType -eq 'colls' -and $Body -eq "{ `"id`": `"$($script:testCollection1)`" }" } `
+                    -ParameterFilter $invokecosmosdbrequest_parameterfilter `
                     -Exactly -Times 1
             }
         }
 
-        Context 'Called with connection parameter and an Id and OfferThroughput parameter' {
+        Context 'Called with context parameter and an Id and OfferThroughput parameter' {
             $script:result = $null
+            $invokecosmosdbrequest_parameterfilter = {
+                $Method -eq 'Post' -and `
+                    $ResourceType -eq 'colls' -and `
+                    $Body -eq (ConvertTo-Json -Depth 10 -InputObject @{ id = $script:testCollection1 } ) -and `
+                    $Headers['x-ms-offer-throughput'] -eq 400
+            }
 
             Mock `
                 -CommandName Invoke-CosmosDbRequest `
-                -ParameterFilter { $Method -eq 'Post' -and $ResourceType -eq 'colls' -and $Body -eq "{ `"id`": `"$($script:testCollection1)`" }" -and $Headers['x-ms-offer-throughput'] -eq 400 } `
+                -ParameterFilter $invokecosmosdbrequest_parameterfilter `
                 -MockWith { ConvertFrom-Json -InputObject $script:testJsonSingle }
 
             It 'Should not throw exception' {
                 $newCosmosDbCollectionParameters = @{
-                    Connection      = $script:testConnection
+                    Context         = $script:testContext
                     Id              = $script:testCollection1
                     OfferThroughput = 400
                 }
@@ -197,24 +314,30 @@ InModuleScope CosmosDB {
             It 'Should call expected mocks' {
                 Assert-MockCalled `
                     -CommandName Invoke-CosmosDbRequest `
-                    -ParameterFilter { $Method -eq 'Post' -and $ResourceType -eq 'colls' -and $Body -eq "{ `"id`": `"$($script:testCollection1)`" }" -and $Headers['x-ms-offer-throughput'] -eq 400 } `
+                    -ParameterFilter $invokecosmosdbrequest_parameterfilter `
                     -Exactly -Times 1
             }
         }
 
-        Context 'Called with connection parameter and an Id and OfferType parameter' {
+        Context 'Called with context parameter and an Id and OfferType parameter' {
             $script:result = $null
+            $invokecosmosdbrequest_parameterfilter = {
+                $Method -eq 'Post' -and `
+                    $ResourceType -eq 'colls' -and `
+                    $Body -eq (ConvertTo-Json -Depth 10 -InputObject @{ id = $script:testCollection1 } ) -and `
+                    $Headers['x-ms-offer-type'] -eq 'S2'
+            }
 
             Mock `
                 -CommandName Invoke-CosmosDbRequest `
-                -ParameterFilter { $Method -eq 'Post' -and $ResourceType -eq 'colls' -and $Body -eq "{ `"id`": `"$($script:testCollection1)`" }" -and $Headers['x-ms-offer-type'] -eq 'S2' } `
+                -ParameterFilter $invokecosmosdbrequest_parameterfilter `
                 -MockWith { ConvertFrom-Json -InputObject $script:testJsonSingle }
 
             It 'Should not throw exception' {
                 $newCosmosDbCollectionParameters = @{
-                    Connection = $script:testConnection
-                    Id         = $script:testCollection1
-                    OfferType  = 'S2'
+                    Context   = $script:testContext
+                    Id        = $script:testCollection1
+                    OfferType = 'S2'
                 }
 
                 { $script:result = New-CosmosDbCollection @newCosmosDbCollectionParameters } | Should -Not -Throw
@@ -227,22 +350,33 @@ InModuleScope CosmosDB {
             It 'Should call expected mocks' {
                 Assert-MockCalled `
                     -CommandName Invoke-CosmosDbRequest `
-                    -ParameterFilter { $Method -eq 'Post' -and $ResourceType -eq 'colls' -and $Body -eq "{ `"id`": `"$($script:testCollection1)`" }" -and $Headers['x-ms-offer-type'] -eq 'S2' } `
+                    -ParameterFilter $invokecosmosdbrequest_parameterfilter `
                     -Exactly -Times 1
             }
         }
 
-        Context 'Called with connection parameter and an Id and PartitionKey parameter' {
+        Context 'Called with context parameter and an Id and PartitionKey parameter' {
             $script:result = $null
+            $invokecosmosdbrequest_parameterfilter = {
+                $Method -eq 'Post' -and `
+                    $ResourceType -eq 'colls' -and `
+                    $Body -eq (ConvertTo-Json -Depth 10 -InputObject @{
+                        id           = $script:testCollection1
+                        partitionKey = @{
+                            paths = @('/partitionkey')
+                            kind  = 'Hash'
+                        }
+                    })
+            }
 
             Mock `
                 -CommandName Invoke-CosmosDbRequest `
-                -ParameterFilter { $Method -eq 'Post' -and $ResourceType -eq 'colls' -and $Body -eq "{ `"id`": `"$($script:testCollection1)`", `"partitionKey`": { `"paths`": [ `"/partitionkey`" ], `"kind`": `"Hash`" } }" } `
+                -ParameterFilter $invokecosmosdbrequest_parameterfilter `
                 -MockWith { ConvertFrom-Json -InputObject $script:testJsonSingle }
 
             It 'Should not throw exception' {
                 $newCosmosDbCollectionParameters = @{
-                    Connection   = $script:testConnection
+                    Context      = $script:testContext
                     Id           = $script:testCollection1
                     PartitionKey = 'partitionkey'
                 }
@@ -257,12 +391,12 @@ InModuleScope CosmosDB {
             It 'Should call expected mocks' {
                 Assert-MockCalled `
                     -CommandName Invoke-CosmosDbRequest `
-                    -ParameterFilter { $Method -eq 'Post' -and $ResourceType -eq 'colls' -and $Body -eq "{ `"id`": `"$($script:testCollection1)`", `"partitionKey`": { `"paths`": [ `"/partitionkey`" ], `"kind`": `"Hash`" } }" } `
+                    -ParameterFilter $invokecosmosdbrequest_parameterfilter `
                     -Exactly -Times 1
             }
         }
 
-        Context 'Called with connection parameter and an Id and OfferType and OfferThrougput' {
+        Context 'Called with context parameter and an Id and OfferType and OfferThrougput' {
             $script:result = $null
 
             Mock `
@@ -271,7 +405,7 @@ InModuleScope CosmosDB {
 
             It 'Should not throw exception' {
                 $newCosmosDbCollectionParameters = @{
-                    Connection      = $script:testConnection
+                    Context         = $script:testContext
                     Id              = $script:testCollection1
                     OfferThroughput = 400
                     OfferType       = 'S1'
@@ -294,7 +428,7 @@ InModuleScope CosmosDB {
             { Get-Command -Name Remove-CosmosDbCollection -ErrorAction Stop } | Should -Not -Throw
         }
 
-        Context 'Called with connection parameter and an Id' {
+        Context 'Called with context parameter and an Id' {
             $script:result = $null
 
             Mock `
@@ -303,8 +437,8 @@ InModuleScope CosmosDB {
 
             It 'Should not throw exception' {
                 $removeCosmosDbCollectionParameters = @{
-                    Connection = $script:testConnection
-                    Id         = $script:testCollection1
+                    Context = $script:testContext
+                    Id      = $script:testCollection1
                 }
 
                 { $script:result = Remove-CosmosDbCollection @removeCosmosDbCollectionParameters } | Should -Not -Throw